requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests